requirements.txt
# TCDD E-Bilet Checker Dependencies

# Core
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Optional - for development
# pytest>=7.0.0
# black>=23.0.0
# mypy>=1.0.0
